Updated to satisfy script analyzer

+++ Helpers/Public/Set-UnixLineEnding.ps1
@@ -1,13 +1,22 @@
 ﻿function Set-UnixLineEnding {
-    [CmdletBinding()]
+    [CmdletBinding(SupportsShouldProcess, ConfirmImpact='High')]
     param(
         [Parameter(Mandatory)]
         [ValidateNotNullOrEmpty()]
         [string[]]
         $Path
     )
-
-    PROCESS {
+    
+    begin {
+        if (-not $PSBoundParameters.ContainsKey('Confirm')) {
+            $ConfirmPreference = $PSCmdlet.SessionState.PSVariable.GetValue('ConfirmPreference')
+        }
+        if (-not $PSBoundParameters.ContainsKey('WhatIf')) {
+            $WhatIfPreference = $PSCmdlet.SessionState.PSVariable.GetValue('WhatIfPreference')
+        }
+    }
+    
+    process {
         foreach ($FilePath in $Path) {
             Write-Verbose ('Processing {0}' -f $FilePath)
 
@@ -21,7 +30,9 @@
             $NormalizedContent = [System.Text.Encoding]::ASCII.GetString($NormalizedBytes)
             #$NormalizedContent
 
-            Set-Content -Value $NormalizedContent -Path $FilePath
+            if ($Force -or $PSCmdlet.ShouldProcess("ShouldProcess?")) {
+                Set-Content -Value $NormalizedContent -Path $FilePath
+            }
         }
     }
 }
